requirements.txt
asgiref~=3.6.0
backports.zoneinfo~=0.2.1
Django~=4.1.7
django-cors-headers~=3.14.0
djangorestframework~=3.14.0
pip~=22.3.1
setuptools~=65.5.1
sqlparse~=0.4.3
tzdata~=2022.7
wheel~=0.38.4
